+++ scripts/call-recording_transcribe_voicebase_url.rb
@@ -9,7 +9,7 @@ require 'pp'
 # Use the rc_config_sample.env.txt file as a scaffold
 
 client = RingCentralSdk::REST::Client.new do |config|
-  config.dotenv = true
+  config.load_env = true
 end
 
 def transcribe_recordings(rcsdk, vbsdk)