+++ scripts/fax_download_new.rb
@@ -7,7 +7,7 @@ require 'pp'
 # Use the rc_config_sample.env.txt file as a scaffold
 
 client = RingCentralSdk::REST::Client.new do |config|
-  config.dotenv = true
+  config.load_env = true
 end
 
 # RcEventFaxDownloader is a helper class to download faxes